Refactored Form to use hooks
